md2docx/requirements.txt
python-docx>=1.1.0
markdown>=3.7
beautifulsoup4>=4.12.0
